% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force algorithm for the knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}
}
\value{
A list with maximum value and weights adding up to the maximum value
}
\description{
Brute force algorithm for the knapsack problem
}
\details{
This algorithm gives all possbile values with good accuracy , and also gives the maximum value for the knapsack
}
\examples{
\dontrun{
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
}
