% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy algorithm for the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}
}
\value{
A list with maximum value and weights adding up to the maximum value
}
\description{
Greedy algorithm for the knapsack problem
}
\details{
This algorithm gives the approximate answer to our problem,it is able to show at least 50% of the true maximum values. And the value given by this algorithm has low computational time.
}
\examples{
\dontrun{
greedy_knapsack(knapsack_objects[1:800], 3500)
greedy_knapsack(knapsack_objects[1:1200], 2000)
}
}
