% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming algorithm for the knapsack problem}
\usage{
knapsack_dynamic(x, W, fast = FALSE)
}
\arguments{
\item{x}{is a data.frame containing variables 'w' and 'v', weights and values}

\item{W}{is the total size of the knapsack}

\item{fast}{is set to TRUE when running Rcpp method to optimize the code}
}
\value{
A list with maximum value and weights adding up to the maximum value
}
\description{
Dynamic programming algorithm for the knapsack problem
}
\details{
An algorithm that can solve the knapsack problem exact by iterating over all possible values of w all possbile values with good accuracy , and also gives the maximum value for the knapsack
}
\examples{
\dontrun{
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
}
}
